refactor(hooks): merge viewport/device sensors into useViewport
Combine useBreakpoint, useIsMobileDevice and useKeyboardInset — three hooks
that all read a live browser source (viewport width / user-agent / on-screen
keyboard) via useSyncExternalStore — into one useViewport module. Co-locating
keeps the deliberate viewport-vs-device distinction visible in one place.

Left untouched on purpose: useMountParam (already pairs with useEffectiveMount
in its own mount-domain file), useCountryCode (geo), and the compare hooks —
merging those would group by name, not cohesion.
