+++ src/components/LensSearchDialog.tsx
@@ -14,7 +14,7 @@ import { Search, X } from "lucide-react";
 import { useTranslations } from "next-intl";
 import { useRouter, Link } from "@/i18n/navigation";
 import { mountToUrlSegment } from "@/lib/mount";
-import { useKeyboardInset } from "@/hooks/useKeyboardInset";
+import { useKeyboardInset } from "@/hooks/useViewport";
 import { buildLensSearchIndex, searchLensIndex } from "@/lib/lens-search";
 import type { Lens } from "@/lib/types";
 import { cn } from "@/lib/utils";
