+++ src/hooks/useShareCapabilities.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { useBreakpoint } from "@/hooks/useBreakpoint";
+import { useBreakpoint } from "@/hooks/useViewport";
 
 export interface ShareCapabilities {
   mounted: boolean;
